Fix TaskStopped exceptions when outbound requests time out (#4690)
