+++ changelog.d/4690.bugfix
@@ -1,0 +1,1 @@
+Fix TaskStopped exceptions in logs when outbound requests time out.
